tests/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "contoso-tests"
version = "0.1.0"
dependencies = [
    "pytest>=8.0",
    "psycopg2-binary>=2.9",
    "redis>=5.0",
    "boto3>=1.34",
]

[tool.pytest.ini_options]
testpaths = ["."]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
